LinkCollector: -also create backup when init fails, now load+init will create backups DownloadController: -also create backup when init fails, now load+init will create backups SecondLevelLaunch: added more headless checks
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@26732 ebf7c1c2-ba36-0410-9fe8-c592906822b4

Former-commit-id: 509331ee6e9dde6746cef3a83c67473910d2f61a

+++ src/jd/controlling/downloadcontroller/DownloadController.java
@@ -438,10 +438,12 @@ public class DownloadController extends PackageController<FilePackage, DownloadL
             protected Void run() throws RuntimeException {
                 logger.info("Init DownloadList");
                 LinkedList<FilePackage> lpackages = null;
+                File loadedList = null;
                 for (File downloadList : findAvailableDownloadLists()) {
                     try {
                         lpackages = load(downloadList);
                         if (lpackages != null) {
+                            loadedList = downloadList;
                             break;
                         }
                     } catch (final Throwable e) {
@@ -479,6 +481,23 @@ public class DownloadController extends PackageController<FilePackage, DownloadL
                     }
                     updateUniqueAlltimeIDMaps(lpackages);
                     eventSender.fireEvent(new DownloadControllerEventStructureRefresh());
+                } catch (final Throwable e) {
+                    if (loadedList != null) {
+                        final File renameTo = new File(loadedList.getAbsolutePath() + ".backup");
+                        boolean backup = false;
+                        try {
+                            if (loadedList.exists()) {
+                                if (loadedList.renameTo(renameTo) == false) {
+                                    IO.copyFile(loadedList, renameTo);
+                                }
+                                backup = true;
+                            }
+                        } catch (final Throwable e2) {
+                            logger.log(e2);
+                        }
+                        logger.severe("Could backup " + loadedList + " to " + renameTo + " ->" + backup);
+                    }
+                    logger.log(e);
                 } finally {
                     DOWNLOADLIST_LOADED.setReached();
                 }
@@ -581,17 +600,19 @@ public class DownloadController extends PackageController<FilePackage, DownloadL
                         zip.close();
                     } catch (final Throwable e2) {
                     }
-                    File renameTo = new File(file.getAbsolutePath() + ".broken");
+                    final File renameTo = new File(file.getAbsolutePath() + ".backup");
                     boolean backup = false;
                     try {
-                        if (file.renameTo(renameTo) == false) {
-                            IO.copyFile(file, renameTo);
+                        if (file.exists()) {
+                            if (file.renameTo(renameTo) == false) {
+                                IO.copyFile(file, renameTo);
+                            }
+                            backup = true;
                         }
-                        backup = true;
                     } catch (final Throwable e2) {
                     }
-                    logger.log(e);
                     logger.severe("Could backup " + file + " to " + renameTo + " ->" + backup);
+                    logger.log(e);
                 } finally {
                     try {
                         zip.close();
